app/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
Flask-CORS==4.0.0
gunicorn==21.2.0
prometheus-client==0.17.1
prometheus-flask-exporter